Add password input (#87)
When using TOTP, a password is required to login. While steamcmd is
perfectly fine with accepting the password as part of the username input
(separated with a space), the way this action is set up, the second
steamcmd invocation where the upload happens would then also use the
password, causing steamcmd to error because the password was provided
without the TOTP code.

Rather than use the TOTP code there again (as proposed in #86), add the
password in the test step (which will cache the credentials), and retain
it from the upload step.

In the configVdf case, the password will be empty, and steamcmd will
ignore it, keeping the previous mechanics.